requirements.txt
requests
datetime
scikit-learn
tensorflow
keras
python-dotenv
pandas
streamlit
matplotlib
seaborn
twilio
